+++ src/main/java/de/alpharogroup/collections/set/SetExtensions.java
@@ -26,9 +26,13 @@ package de.alpharogroup.collections.set;
 
 import java.util.Collection;
 import java.util.Collections;
+import java.util.Comparator;
 import java.util.HashSet;
 import java.util.Set;
+import java.util.SortedSet;
+import java.util.TreeSet;
 
+import de.alpharogroup.collections.CollectionExtensions;
 import lombok.experimental.UtilityClass;
 
 /**
@@ -39,13 +43,13 @@ public class SetExtensions
 {
 
 	/**
-	 * Factory method for create new {@link HashSet} and returns as {@link Set}.
+	 * Factory method for create new {@link HashSet} and will be returned as {@link Set}
 	 *
 	 * @param <T>
 	 *            the generic type of the elements
 	 * @param elements
-	 *            the elements to add in the new {@link HashSet}.
-	 * @return the new {@link HashSet} and returns as {@link Set}.
+	 *            the elements to add in the new {@link HashSet}
+	 * @return the new {@link HashSet}
 	 */
 	@SafeVarargs
 	public static final <T> Set<T> newHashSet(final T... elements)
@@ -54,21 +58,21 @@ public class SetExtensions
 	}
 
 	/**
-	 * Factory method for create new {@link HashSet} and returns as {@link Set}.
+	 * Factory method for create new {@link HashSet} and will be returned as {@link Set}
 	 *
 	 * @param <T>
 	 *            the generic type of the elements
 	 * @param collection
 	 *            the optional collection that will be added to the new list
 	 * @param elements
-	 *            the elements to add in the new {@link HashSet}.
-	 * @return the new {@link HashSet} and returns as {@link Set}.
+	 *            the elements to add in the new {@link HashSet}
+	 * @return the new {@link HashSet}
 	 */
 	@SafeVarargs
 	public static final <T> Set<T> newHashSet(final Collection<T> collection, final T... elements)
 	{
 		final Set<T> set;
-		if (collection != null && !collection.isEmpty())
+		if (CollectionExtensions.isNotEmpty(collection))
 		{
 			set = new HashSet<>(collection);
 		}
@@ -84,31 +88,107 @@ public class SetExtensions
 	}
 
 	/**
-	 * Checks if a Set is null or empty.
+	 * Factory method for create new {@link HashSet} and will be returned as {@link Set}
 	 *
 	 * @param <T>
 	 *            the generic type of the elements
-	 * @param set
-	 *            The Set to check.
-	 * @return true if the set is null or empty otherwise false.
+	 * @param elements
+	 *            the elements to add in the new {@link HashSet}
+	 * @return the new {@link HashSet}
 	 */
-	public static <T> boolean isEmpty(final Set<T> set)
+	@SafeVarargs
+	public static final <T> SortedSet<T> newTreeSet(final T... elements)
 	{
-		return set == null || set.isEmpty();
+		return newTreeSet(null, elements);
 	}
 
 	/**
-	 * Checks if the given Set is not null or empty.
+	 * Factory method for create new {@link TreeSet} and will be returned as {@link SortedSet}
+	 *
+	 * @param <T>
+	 *            the generic type of the elements
+	 * @param collection
+	 *            the optional collection that will be added to the new list
+	 * @param elements
+	 *            the elements to add in the new {@link HashSet}
+	 * @return the new {@link HashSet}
+	 */
+	@SafeVarargs
+	public static final <T> SortedSet<T> newTreeSet(final Collection<T> collection,
+		final T... elements)
+	{
+		return newTreeSet(collection, null, elements);
+	}
+
+	/**
+	 * Factory method for create new {@link TreeSet} and will be returned as {@link SortedSet}
+	 *
+	 * @param <T>
+	 *            the generic type of the elements
+	 * @param collection
+	 *            the optional collection that will be added to the new list
+	 * @param comparator
+	 *            the comparator
+	 * @param elements
+	 *            the elements to add in the new {@link HashSet}
+	 * @return the new {@link HashSet}
+	 */
+	@SafeVarargs
+	public static final <T> SortedSet<T> newTreeSet(final Collection<T> collection,
+		final Comparator<T> comparator, final T... elements)
+	{
+		final SortedSet<T> sortedSet;
+		if (comparator != null)
+		{
+			sortedSet = new TreeSet<>(comparator);
+		}
+		else
+		{
+			sortedSet = new TreeSet<>();
+		}
+		if (CollectionExtensions.isNotEmpty(collection))
+		{
+			sortedSet.addAll(collection);
+		}
+		if (0 < elements.length)
+		{
+			Collections.addAll(sortedSet, elements);
+		}
+		return sortedSet;
+	}
+
+	/**
+	 * Checks if a Set is null or empty
 	 *
 	 * @param <T>
 	 *            the generic type of the elements
 	 * @param set
-	 *            The Set to check.
-	 * @return true if the set is null or empty otherwise false.
+	 *            The Set to check
+	 * @return true if the set is null or empty otherwise false
+	 * @deprecated use instead the same name method in the class CollectionExtensions. Note: will be
+	 *             removed in the next minor release
 	 */
+	@Deprecated
+	public static <T> boolean isEmpty(final Set<T> set)
+	{
+		return CollectionExtensions.isEmpty(set);
+	}
+
+	/**
+	 * Checks if the given Set is not null or empty
+	 *
+	 * @param <T>
+	 *            the generic type of the elements
+	 * @param set
+	 *            The Set to check
+	 * @return true if the set is null or empty otherwise false.
+	 * @deprecated use instead the same name method in the class CollectionExtensions. Note: will be
+	 *             removed in the next minor release
+	 */
+	@Deprecated
 	public static <T> boolean isNotEmpty(final Set<T> set)
 	{
-		return set != null && !set.isEmpty();
+		return CollectionExtensions.isNotEmpty(set);
 	}
 
 	/**
@@ -123,6 +203,20 @@ public class SetExtensions
 	public static <T> Set<T> toSet(final Collection<T> collection)
 	{
 		return newHashSet(collection);
+	}
+
+	/**
+	 * Converts the given {@link Collection} to a {@link SortedSet}.
+	 *
+	 * @param <T>
+	 *            the generic type of the elements
+	 * @param collection
+	 *            the collection
+	 * @return A new {@link SortedSet}
+	 */
+	public static <T> SortedSet<T> toSortedSet(final Collection<T> collection)
+	{
+		return newTreeSet(collection);
 	}
 
 }